finally removed all filtering from the BrowseContent. In terms of filtering, BrowseContent is now a DumbComponent passing data to others or waiting for filtered data but not actually dealing with the filters themself. FIlter logic contained within filterOperations.ts and FilterSection
